Fix: Stalfos crash in Tower Collapse 3
Fixes a crash caused by defeating the Stalfos required to free Zelda in Tower Collapse 3.

The previous Stalfos change caused individual pieces to ignore wall collision. A new restriction was put up to check if the pieces wouldn't fall through the floor near walls. This is done by comparing the height of the parent actor, which itself causes said crash. Checking if the parent isn't NULL beforehand fixes this.